User_Append.py
This personal project objective is to create a system where user input is stored into a database. This database acts as a social media platform where user's comments can be stored up to a certain amount per time period. I will be putting to practice what I've learn so far in python in order to create a system that's able to do the 8 listed steps in the beginning of the program.